requirements.txt
httplib2==0.9.2
matplotlib==3.0.2
numpy==1.16.2
opencv-contrib-python
opencv-python
Pillow==6.0.0
scipy==1.2.1
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
torch==1.0.1.post2
torchvision==0.2.2.post3
